finish: Project1 StandaloneKV

+++ kv/util/engine_util/engines.go
@@ -41,8 +41,10 @@ func (en *Engines) Close() error {
 	if err := en.Kv.Close(); err != nil {
 		return err
 	}
-	if err := en.Raft.Close(); err != nil {
-		return err
+	if en.Raft != nil {
+		if err := en.Raft.Close(); err != nil {
+			return err
+		}
 	}
 	return nil
 }
@@ -54,8 +56,10 @@ func (en *Engines) Destroy() error {
 	if err := os.RemoveAll(en.KvPath); err != nil {
 		return err
 	}
-	if err := os.RemoveAll(en.RaftPath); err != nil {
-		return err
+	if en.Raft != nil {
+		if err := os.RemoveAll(en.RaftPath); err != nil {
+			return err
+		}
 	}
 	return nil
 }
